Extract Reports::Base parent class
