+++ app/models/reports/daily.rb
@@ -3,19 +3,8 @@
 module Reports
   # Builds a snapshot of the job offers an administrator is involved in:
   # offers published the previous day plus open applications grouped by state.
-  class Daily
+  class Daily < Base
     include Rails.application.routes.url_helpers
-
-    Section = Struct.new(:key, :human_state, :count, :items, keyword_init: true)
-    Item = Struct.new(:title, :link, keyword_init: true)
-
-    def initialize(administrator)
-      @administrator = administrator
-    end
-
-    def sections
-      @sections ||= [new_offers_section] + JobApplication::ORDERED_STATES.map { |state| applications_section(state) }
-    end
 
     private
